+++ .travis.yml
@@ -25,7 +25,7 @@ env:
     - TESTENV=py35-trial
     - TESTENV=py27-nobyte
     - TESTENV=doctesting
-    - TESTENV=py27-cxfreeze
+    - TESTENV=freeze
 
 script: tox --recreate -e $TESTENV
