--- testing/cx_freeze/runtests_setup.py
@@ -1,15 +1,0 @@
-"""
-Sample setup.py script that generates an executable with pytest runner embedded.
-"""
-if __name__ == '__main__':
-    from cx_Freeze import setup, Executable
-    import pytest
-
-    setup(
-        name="runtests",
-        version="0.1",
-        description="example of how embedding py.test into an executable using cx_freeze",
-        executables=[Executable("runtests_script.py")],
-        options={"build_exe": {'includes': pytest.freeze_includes()}},
-    )
-
