--- testing/cx_freeze/tests/test_trivial.py
@@ -1,6 +1,0 @@
-
-def test_upper():
-    assert 'foo'.upper() == 'FOO'
-
-def test_lower():
-    assert 'FOO'.lower() == 'foo'
