+++ testing/freeze/tests/test_trivial.py
@@ -1,0 +1,6 @@
+
+def test_upper():
+    assert 'foo'.upper() == 'FOO'
+
+def test_lower():
+    assert 'FOO'.lower() == 'foo'
