+++ tox.ini
@@ -5,7 +5,7 @@ distshare={homedir}/.tox/distshare
 envlist=
      linting,py26,py27,py33,py34,py35,pypy,
      {py27,py35}-{pexpect,xdist,trial},
-     py27-nobyte,doctesting,py27-cxfreeze
+     py27-nobyte,doctesting,freeze
 
 [testenv]
 commands= py.test --lsof -rfsxX {posargs:testing}
@@ -124,12 +124,11 @@ changedir=testing
 commands=
     {envpython} {envbindir}/py.test-jython -rfsxX {posargs}
 
-[testenv:py27-cxfreeze]
-changedir=testing/cx_freeze
-platform=linux|darwin
+[testenv:freeze]
+changedir=testing/freeze
+deps=pyinstaller
 commands=
-    {envpython} install_cx_freeze.py
-    {envpython} runtests_setup.py build --build-exe build
+    {envpython} create_executable.py
     {envpython} tox_run.py
